Rename FinalAffiliationField file to AffiliationField
