+++ indico/web/client/js/react/components/AffiliationField.jsx
