+++ indico/web/client/js/react/components/index.js
@@ -63,5 +63,5 @@ export {default as WTFSearchDropdown, SearchDropdown} from './WTFSearchDropdown'
 export {default as ICSCalendarLink} from './ICSCalendarLink';
 export {default as Favorite} from './Favorite';
 export {default as PersonDetailsModal} from './PersonDetailsModal';
-export {default as FinalAffiliationField} from './FinalAffiliationField';
+export {default as FinalAffiliationField} from './AffiliationField';
 export {default as Captcha} from './Captcha';
